+++ src/globals.ts
@@ -1,0 +1,10 @@
+import {Config} from './config';
+
+let configuration: Config | undefined;
+const bus = new EventTarget();
+
+function set_configuration(cfg: Config) {
+  configuration = cfg;
+}
+
+export {configuration, bus, set_configuration, Config};
